requirements.txt
tensorflow==2.8
protobuf==3.19.0
flask 
pillow 
numpy